medical_segmentation/requirements.txt
boto==2.49.0
split-folders==0.4.3
pandas==1.2.4
tensorflow==2.5.0
scipy==1.6.3
numpy==1.19.5
ipython==7.22.0
jupyter==1.0.0
matplotlib==3.4.1
keras-unet
Pillow==8.2.0
click==8.0.1
pydicom==2.1.2
ctg-surface-distance==0.1
